norm: create execute_pipe_command.c
